[tvzavr] My setup [also problem]
- This setup imitates setup in my application.
Deeplinks are handled by translucent activity which tries to get af_dp out of incoming OneLink and then passes this data to other activities to react.

Testing:

- remove the app from recent tasks (if is is running)
- open app with deep link: https://androidsampleapp.onelink.me/Pvqj
- AppsFlyerConversionListener.onAppAttribution() is called
- press Home
- open app with launcher icon
- AppsFlyerConversionListener.onAppAttribution() IS called again (why?)

Sometimes onAppAttribution() is not called the first time you re-open the app with launcher icon, but it always is called again if you keep re-opening the app.